safari-600.7-branch/LayoutTests/fast/frames/adopt-object-into-itself-expected.txt

ALERT: Error: HierarchyRequestError: DOM Exception 3
